+++ packages/create-tutorial/CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.0.3](https://github.com/stackblitz/tutorialkit/compare/0.2.0...0.0.3) "create-tutorial" (2024-08-30)
+
+
+
+## [0.0.3](https://github.com/stackblitz/tutorialkit/compare/0.1.6...0.0.3) "create-tutorial" (2024-08-28)
+
+
+
 ## [0.0.3](https://github.com/stackblitz/tutorialkit/compare/0.1.5...0.0.3) "create-tutorial" (2024-08-26)